move existing debug stuff to folder
